Revert "pqx: add -pqx.c flag for debugging test database just before cleanup"
