--- .github/actions/ci/test.sh
@@ -1,1 +1,0 @@
-docker run -it -v /Users/bmizerany/src/tier/:/github/workspace $1 bash
