--- .github/workflows/ci.yml
@@ -1,12 +1,0 @@
-name: test
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: ./.github/actions/ci
